+++ CHANGELOG.md
@@ -5,8 +5,10 @@
   - Section 4: DOM Projects
     - Number Guesser [Part 2] - Validation & Winning Case
     - Number Guesser [Part 3] - Lose Case & Game Over
+    - Number Guesser [Part 4] - Play Again
 - Validation logic and logic for the winning case was added.
 - Logic for lose case and game over logic was added.
+- Logic for playing again was added.
 
 ## August 26, 2020
 - Modern JavaScript From The Beginning (Udemy)